Fix bugs and make use of better APIs
code makes use of buildHttpMessage now which should take care of HTTP/2 aswell. Also removed the need for manually updating the CL, which was pretty ugly, urgh.

Further checks now if we have the academyLabHeader div in the HTML, in order to not attempt to transform other HTML responses which would cause weird responses (i.e HTTP 5xx's )
and sometimes broke things.